Made variable declaration possible. Now variables that correctly declared are stored in special array and each of them is represented as an object, which hold name, class(array or no), type, value and length for arrays
Added check at the beginning of the chechSyntax function. It controls that "VAR" has to be the first one, if there's no "VAR" clause -- only empty BEGIN/END brackets allowed. That's what "onlyBeginEnd" checks.
Also added func "checkVar" where we identify new variables and check if VAR statement is written correctly.
There we have "temporary vault" where we put just found variables.
Also created expressionsTable, which is an an analog of lexemsTable, but instead of lexems we have that expressions that occur in the input string